+++ src/main/java/requestModel/ImportRequestModel.java
@@ -1,6 +1,6 @@
 package requestModel;
 
-import useCaseInteractor.IcsParser;
+import useCaseInteractor.Import.IcsParser;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
