--- src/main/java/useCaseInteractor/AddTask.java
@@ -1,4 +1,0 @@
-package useCaseInteractor;
-
-public class AddTask {
-}
