--- src/main/java/useCaseInteractor/DeleteTask.java
@@ -1,4 +1,0 @@
-package useCaseInteractor;
-
-public class DeleteTask {
-}
