--- src/main/java/useCaseInteractor/EditScheduleItem.java
@@ -1,4 +1,0 @@
-package useCaseInteractor;
-
-public class EditScheduleItem {
-}
