--- src/main/java/useCaseInteractor/EditTask.java
@@ -1,4 +1,0 @@
-package useCaseInteractor;
-
-public class EditTask {
-}
