+++ src/main/java/useCaseInteractor/Import/IcsParser.java
@@ -1,4 +1,4 @@
-package useCaseInteractor;
+package useCaseInteractor.Import;
 
 import net.fortuna.ical4j.data.CalendarBuilder;
 import net.fortuna.ical4j.data.ParserException;
