+++ src/main/java/useCaseInteractor/Schedule/DeleteScheduleItem.java
@@ -1,4 +1,4 @@
-package useCaseInteractor;
+package useCaseInteractor.Schedule;
 
 import boundary.DeleteScheduleItemInputBoundary;
 import requestModel.ScheduleItemRequestModel;
