+++ src/main/java/useCaseInteractor/Schedule/EditScheduleItem.java
@@ -1,0 +1,4 @@
+package useCaseInteractor.Schedule;
+
+public class EditScheduleItem {
+}
