+++ src/main/java/useCaseInteractor/Task/AddTask.java
@@ -1,0 +1,4 @@
+package useCaseInteractor.Task;
+
+public class AddTask {
+}
