+++ src/main/java/useCaseInteractor/Task/DeleteTask.java
@@ -1,0 +1,4 @@
+package useCaseInteractor.Task;
+
+public class DeleteTask {
+}
