+++ src/main/java/useCaseInteractor/Task/EditTask.java
@@ -1,0 +1,4 @@
+package useCaseInteractor.Task;
+
+public class EditTask {
+}
