+++ src/main/java/useCaseInteractor/User/createAccountForm.java
@@ -1,4 +1,4 @@
-package useCaseInteractor;
+package useCaseInteractor.User;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
