+++ src/main/java/useCaseInteractor/User/setUsername.java
@@ -1,4 +1,4 @@
-package useCaseInteractor;
+package useCaseInteractor.User;
 
 import controller.TimetableController;
