+++ src/main/java/useCaseInteractor/User/userCollection.java
@@ -1,4 +1,4 @@
-package useCaseInteractor;
+package useCaseInteractor.User;
 
 import entity.User;
